Move TableIterator to values package
