+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/TableProvider.java
@@ -27,6 +27,7 @@ import org.ballerinalang.jvm.util.exceptions.BallerinaException;
 import org.ballerinalang.jvm.values.ArrayValue;
 import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.RefValue;
+import org.ballerinalang.jvm.values.TableIterator;
 
 import java.io.ByteArrayInputStream;
 import java.sql.Connection;
