+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/TableValue.java
@@ -19,7 +19,6 @@ package org.ballerinalang.jvm.values;
 
 import org.ballerinalang.jvm.ColumnDefinition;
 import org.ballerinalang.jvm.DataIterator;
-import org.ballerinalang.jvm.TableIterator;
 import org.ballerinalang.jvm.TableProvider;
 import org.ballerinalang.jvm.TableUtils;
 import org.ballerinalang.jvm.types.BStructureType;
